# source/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/v1/booking": {
            "post": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "createBooking",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/booking/extend": {
            "put": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "extendStay",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExtendStayInput"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "tags": [
                    "booking-controller"
                ],
                "operationId": "helloWorld",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HealthResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookingInput": {
                "type": "object",
                "properties": {
                    "guestName": {
                        "type": "string"
                    },
                    "unitID": {
                        "type": "string"
                    },
                    "checkInDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "numberOfNights": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ExtendStayInput": {
                "type": "object",
                "properties": {
                    "guestName": {
                        "type": "string"
                    },
                    "unitID": {
                        "type": "string"
                    },
                    "additionalNights": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "HealthResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            }
        }
    }
}